refactor(scanner/interactionScreen): realign concerns
move some state into scanner component
move common bits into interaction screen (to be reused for BLE/NFC)
use a promise to signal errors for "local notifications"
